Added initial version of Carousel Site Template
All content is formatted by structures. Currently, the three-column
layout for the heading posts is created by using nested portlets,
not a unique layout. Some text styling is set in the featurette
template and might override theme styling. Future work would be to
add an option to include text in the carousel.