Added Jumbotron site template
Includes two structures: jumbotron heading and heading post.
Jumbotron heading is the same as the structure from the jumbtron
(narrow) site template, just with fewer styling rules. Header post
is the same as the structure in the carousel site template, but
without the image option. Background color is specified in the
jumbotron heading template and may override theme settings. Text is
left unstyled for both structures -- can be specified in the theme.